backend/requirements.txt
fastapi
uvicorn
sympy
python-multipart
requests
transformers
torch
torchvision
pillow
